Revert "[Add] Orders manager layout (#852)"
This reverts commit d981103d1352bb69c548736156d0e028a2893254.
