+++ src/components/Orders/DriversBusyStatusFilter/styles.js
@@ -1,17 +1,11 @@
 import styled, { css } from 'styled-components'
 
 export const Container = styled.div`
-  display: flex;
+  dispaly: flex;
   button {
+    margin: 10px;
     white-space: nowrap;
     font-size: 14px;
-    margin: 10px 0;
-
-    ${props => props.theme?.rtl ? css`
-      margin-left: 10px;
-    ` : css`
-      margin-right: 10px;
-    `}
 
     svg {
       font-size: 20px;
