+++ src/components/Orders/OrdersLateralBar/index.js
@@ -3,7 +3,7 @@ import { useLanguage, useUtils } from 'ordering-components-admin'
 import { useWindowSize } from '../../../hooks/useWindowSize'
 import { XLg, ArrowsAngleExpand, ArrowsAngleContract } from 'react-bootstrap-icons'
 import { useTheme } from 'styled-components'
-import { OriginalOrdersManager } from '../OrdersManager/Layouts/OriginalOrdersManager'
+import { OrdersManager } from '../OrdersManager'
 import { IconButton } from '../../../styles/Buttons'
 
 import {
@@ -141,7 +141,7 @@ export const OrdersLateralBar = (props) => {
             </p>
           </Name>
         </Info>
-        <OriginalOrdersManager
+        <OrdersManager
           isSelectedOrders
           driverId={isDriver ? user.id : null}
           customerId={isCustomer ? user.id : null}
