+++ template/theme.json
@@ -56,12 +56,5 @@
   },
   "defaultLanguages": {
     "LOADING_DELICIOUS_FOOD": "Loading Ordering Dashboard..."
-  },
-  "orders_manager_view": {
-    "components": {
-      "layout": {
-        "type": "original"
-      }
-    }
   }
 }
